refactor: move gitlab to model
