--- src/interfaces/mod.rs
@@ -1,3 +1,0 @@
-// Copyright 2024 bmc::labs GmbH. All rights reserved.
-
-pub use interfaces::gitlab;
